Refactor weapons
