+++ entities/weapons/Weapon.js
