+++ utils/Reward-phase.js
@@ -1,8 +1,8 @@
 import { BOARD_HEIGHT, BOARD_WIDTH, player } from './config.js';
 import { config } from '../take.js';
-import Ezq from '../entities/EZq.js';
-import AOE from '../entities/AOE.js';
-import Freeze from '../entities/Freeze.js';
+import Ezq from '../entities/weapons/EZq.js';
+import AOE from '../entities/weapons/AOE.js';
+import Freeze from '../entities/weapons/Freeze.js';
 
 let numberOfReward;
 let rewardArray;
